Add checking of credit scores and credit reports to month change checklist
Previously, I've been doing this on an *ad hoc* basis multiple times a
month. But now that the situation has been stable for a while and I'm
not looking to make changes in the near future, it is time to stop
having to actively think about it and just limit to doing it monthly
as part of the checklist.